- Now painting tree will generate trees of random sizes. (starting size is also more proportionnal with the character)
git-svn-id: http://svn.code.sf.net/p/irrrpgbuilder/code@30 528fbfa9-60f6-4f4a-a82b-d133dc65ca33

+++ src/Vegetation.cpp
@@ -25,7 +25,7 @@ Vegetation::Vegetation()
     s32 texture = rand()%4;
     ITexture* itexture = NULL;
 
-    size = (f32)0.8;
+    size = 0.8f;
 
     switch(texture)
     {
